Backend/requirements.txt
Flask
flask-cors
flask-pymongo
python-dotenv
bcrypt
PyJWT
psutil
langchain_groq
langchain_core
langchain_community
langchain_huggingface
pymongo
# error_handler
faiss-cpu
gunicorn 
nltk
flask-cors
Authlib